# api_gateway/api/docs/swagger.yaml
definitions:
...
  models.ResponseError:
    properties:
      error: {}
    type: object
  models.ResponseSuccess:
    properties:
      data: {}
      metadata: {}
    type: object
  task_service.CreateTask:
    properties:
      deadline:
        type: string
      task_description:
        type: string
      task_status:
        type: string
      title:
        type: string
      user_id:
        type: string
    type: object
  task_service.TaskChangeStatus:
    properties:
      new_status:
        type: string
      task_id:
        type: string
    type: object
  task_service.UpdateTask:
    properties:
      deadline:
        type: string
      id:
        type: string
      task_description:
        type: number
      title:
        type: string
      user_id:
        type: string
    type: object
...
info:
  contact: {}
  description: This is a To_do_list celler server.
  title: Swagger To_do_list API
  version: "1.0"
paths:
...
  /v1/task/change_status:
    patch:
      consumes:
      - application/json
      description: API for Updating taskes
      parameters:
      - description: Authorization token
        in: header
        name: Authorization
        required: true
        type: string
      - description: task
        in: body
        name: task
        required: true
        schema:
          $ref: '#/definitions/task_service.TaskChangeStatus'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResponseSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ResponseError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      security:
      - ApiKeyAuth: []
      summary: Update task
      tags:
      - task
  /v1/task/create:
    post:
      consumes:
      - application/json
      description: API for creating taskes
      parameters:
      - description: Authorization token
        in: header
        name: Authorization
        required: true
        type: string
      - description: task
        in: body
        name: task
        required: true
        schema:
          $ref: '#/definitions/task_service.CreateTask'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResponseSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ResponseError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      security:
      - ApiKeyAuth: []
      summary: Create task
      tags:
      - task
  /v1/task/delete/{id}:
    delete:
      consumes:
      - application/json
      description: API for deleting task
      parameters:
      - description: Authorization token
        in: header
        name: Authorization
        required: true
        type: string
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResponseSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ResponseError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      security:
      - ApiKeyAuth: []
      summary: Delete task
      tags:
      - task
  /v1/task/get/{id}:
    get:
      consumes:
      - application/json
      description: API for getting task
      parameters:
      - description: Authorization token
        in: header
        name: Authorization
        required: true
        type: string
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResponseSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ResponseError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      security:
      - ApiKeyAuth: []
      summary: Get task
      tags:
      - task
  /v1/task/get_by_task_id/{id}:
    get:
      consumes:
      - application/json
      description: API for getting task
      parameters:
      - description: Authorization token
        in: header
        name: Authorization
        required: true
        type: string
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResponseSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ResponseError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      security:
      - ApiKeyAuth: []
      summary: Get task
      tags:
      - task
  /v1/task/getall:
    get:
      consumes:
      - application/json
      description: API for getting all taskes
      parameters:
      - description: Authorization token
        in: header
        name: Authorization
        required: true
        type: string
      - description: user_id
        in: query
        name: user_id
        required: true
        type: string
      - description: search
        in: query
        name: search
        type: string
      - description: page
        in: query
        name: page
        type: integer
      - description: limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResponseSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ResponseError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      security:
      - ApiKeyAuth: []
      summary: Get all taskes
      tags:
      - task
  /v1/task/update:
    put:
      consumes:
      - application/json
      description: API for Updating tasks
      parameters:
      - description: Authorization token
        in: header
        name: Authorization
        required: true
        type: string
      - description: task
        in: body
        name: task
        required: true
        schema:
          $ref: '#/definitions/task_service.UpdateTask'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResponseSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ResponseError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      security:
      - ApiKeyAuth: []
      summary: Update task
      tags:
      - task
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"